Added test cases for rancher-compose up options
