+++ tests/validation/cattlevalidationtest/core/conftest.py
@@ -19,12 +19,13 @@ def cleanup(super_client):
         except AttributeError:
             pass
     delete_all(super_client, to_delete_env)
-
     to_delete = []
     for i in super_client.list_instance(state='running'):
         try:
             if instance_name_format.match(i.name) or \
-                    instance_name_format_for_services.match(i.name):
+                    instance_name_format_for_services.match(i.name) or \
+                    i.name.startswith("socat-test") or \
+                    i.name.startswith("rancher-compose"):
                 to_delete.append(i)
         except AttributeError:
             pass
