+++ tests/validation/cattlevalidationtest/core/test_dns_services.py
@@ -1,65 +1,6 @@
 from common_fixtures import *  # NOQA
 
-WEB_IMAGE_UUID = "docker:sangeetha/testlbsd:latest"
-SSH_IMAGE_UUID = "docker:sangeetha/testclient:latest"
-
 logger = logging.getLogger(__name__)
-
-
-def create_env_with_2_svc_dns(client, scale_svc, scale_consumed_svc, port):
-
-    launch_config_svc = {"imageUuid": SSH_IMAGE_UUID,
-                         "ports": [port+":22/tcp"]}
-
-    launch_config_consumed_svc = {"imageUuid": WEB_IMAGE_UUID}
-
-    # Create Environment
-    random_name = random_str()
-    env_name = random_name.replace("-", "")
-    env = client.create_environment(name=env_name)
-    env = client.wait_success(env)
-    assert env.state == "active"
-
-    # Create Service
-    random_name = random_str()
-    service_name = random_name.replace("-", "")
-    service = client.create_service(name=service_name,
-                                    environmentId=env.id,
-                                    launchConfig=launch_config_svc,
-                                    scale=scale_svc)
-
-    service = client.wait_success(service)
-    assert service.state == "inactive"
-
-    # Create Consumed Service
-    random_name = random_str()
-    service_name = random_name.replace("-", "")
-
-    consumed_service = client.create_service(
-        name=service_name, environmentId=env.id,
-        launchConfig=launch_config_consumed_svc, scale=scale_consumed_svc)
-
-    consumed_service = client.wait_success(consumed_service)
-    assert consumed_service.state == "inactive"
-
-    # Create Consumed Service
-    random_name = random_str()
-    service_name = random_name.replace("-", "")
-
-    consumed_service1 = client.create_service(
-        name=service_name, environmentId=env.id,
-        launchConfig=launch_config_consumed_svc, scale=scale_consumed_svc)
-
-    consumed_service1 = client.wait_success(consumed_service1)
-    assert consumed_service1.state == "inactive"
-
-    # Create DNS service
-
-    dns = client.create_dnsService(name='WEB1',
-                                   environmentId=env.id)
-    dns = client.wait_success(dns)
-
-    return env, service, consumed_service, consumed_service1, dns
 
 
 def create_environment_with_dns_services(super_client, client,
@@ -850,129 +791,3 @@ def test_dns_add_remove_servicelinks_using_set(super_client, client):
     validate_dns_service(super_client, service, [consumed_service1], port,
                          dns.name)
     delete_all(client, [env])
-
-
-def check_service_map(super_client, service, instance, state):
-    instance_service_map = super_client.\
-        list_serviceExposeMap(serviceId=service.id, instanceId=instance.id,
-                              state=state)
-    assert len(instance_service_map) == 1
-
-
-def get_container_names_list(super_client, env, services):
-    container_names = []
-    for service in services:
-        containers = get_service_container_list(super_client, service)
-        for c in containers:
-            if c.state == "running":
-                container_names.append(c.externalId[:12])
-    return container_names
-
-
-def validate_add_service_link(super_client, service, consumedService):
-    service_maps = super_client. \
-        list_serviceConsumeMap(serviceId=service.id,
-                               consumedServiceId=consumedService.id)
-    assert len(service_maps) == 1
-    service_map = service_maps[0]
-    wait_for_condition(
-        super_client, service_map,
-        lambda x: x.state == "active",
-        lambda x: 'State is: ' + x.state)
-
-
-def validate_remove_service_link(super_client, service, consumedService):
-    service_maps = super_client. \
-        list_serviceConsumeMap(serviceId=service.id,
-                               consumedServiceId=consumedService.id)
-    assert len(service_maps) == 1
-    service_map = service_maps[0]
-    wait_for_condition(
-        super_client, service_map,
-        lambda x: x.state == "removed",
-        lambda x: 'State is: ' + x.state)
-
-
-def get_service_container_list(super_client, service):
-
-    container = []
-    instance_maps = super_client.list_serviceExposeMap(serviceId=service.id,
-                                                       state="active")
-    for instance_map in instance_maps:
-        c = super_client.by_id('container', instance_map.instanceId)
-        containers = super_client.list_container(
-            externalId=c.externalId,
-            include="hosts")
-        assert len(containers) == 1
-        container.append(containers[0])
-
-    return container
-
-
-def validate_dns_service(super_client, service, consumed_services,
-                         exposed_port, dnsname, exclude_instance=None,
-                         exclude_instance_purged=False):
-    time.sleep(5)
-
-    service_containers = get_service_container_list(super_client, service)
-    assert len(service_containers) == service.scale
-
-    for con in service_containers:
-        host = super_client.by_id('host', con.hosts[0].id)
-
-        containers = []
-        expected_dns_list = []
-        expected_link_response = []
-        dns_response = []
-
-        for consumed_service in consumed_services:
-            cons = get_service_container_list(super_client, consumed_service)
-            if exclude_instance_purged:
-                assert len(cons) == consumed_service.scale - 1
-            else:
-                assert len(cons) == consumed_service.scale
-            containers = containers + cons
-        for con in containers:
-            if (exclude_instance is not None) \
-                    and (con.id == exclude_instance.id):
-                logger.info("Excluded from DNS and wget list:" + con.name)
-            else:
-                expected_dns_list.append(con.primaryIpAddress)
-                expected_link_response.append(con.externalId[:12])
-
-        logger.info("Expected dig response List" + str(expected_dns_list))
-        logger.info("Expected wget response List" +
-                    str(expected_link_response))
-
-        # Validate port mapping
-        ssh = paramiko.SSHClient()
-        ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
-        ssh.connect(host.ipAddresses()[0].address, username="root",
-                    password="root", port=int(exposed_port))
-
-        # Validate link containers
-        cmd = "wget -O result.txt http://" + dnsname + \
-              ":80/name.html;cat result.txt"
-        logger.info(cmd)
-        stdin, stdout, stderr = ssh.exec_command(cmd)
-
-        response = stdout.readlines()
-        assert len(response) == 1
-        resp = response[0].strip("\n")
-        logger.info("Actual wget Response" + str(resp))
-        assert resp in (expected_link_response)
-
-        # Validate DNS resolution using dig
-        cmd = "dig " + dnsname + " +short"
-        logger.info(cmd)
-        stdin, stdout, stderr = ssh.exec_command(cmd)
-
-        response = stdout.readlines()
-        logger.info("Actual dig Response" + str(response))
-        assert len(response) == len(expected_dns_list)
-
-        for resp in response:
-            dns_response.append(resp.strip("\n"))
-
-        for address in expected_dns_list:
-            assert address in dns_response
