+++ tests/validation/cattlevalidationtest/core/test_ext_services_links.py
@@ -1,58 +1,5 @@
 from common_fixtures import *  # NOQA
-
-WEB_IMAGE_UUID = "docker:sangeetha/testlbsd:latest"
-SSH_IMAGE_UUID = "docker:sangeetha/testclient:latest"
-
 logger = logging.getLogger(__name__)
-
-
-def create_env_with_ext_svc(client, scale_svc, port):
-
-    launch_config_svc = {"imageUuid": SSH_IMAGE_UUID,
-                         "ports": [port+":22/tcp"]}
-
-    # Create Environment
-    random_name = random_str()
-    env_name = random_name.replace("-", "")
-    env = client.create_environment(name=env_name)
-    env = client.wait_success(env)
-    assert env.state == "active"
-
-    # Create Service
-    random_name = random_str()
-    service_name = random_name.replace("-", "")
-    service = client.create_service(name=service_name,
-                                    environmentId=env.id,
-                                    launchConfig=launch_config_svc,
-                                    scale=scale_svc)
-
-    service = client.wait_success(service)
-    assert service.state == "inactive"
-
-    # Create 2 containers which would be the applications that need to be
-    # serviced by the external service
-
-    c1 = client.create_container(name=random_str(), imageUuid=WEB_IMAGE_UUID)
-    c2 = client.create_container(name=random_str(), imageUuid=WEB_IMAGE_UUID)
-
-    c1 = client.wait_success(c1, 120)
-    assert c1.state == "running"
-    c2 = client.wait_success(c2, 120)
-    assert c2.state == "running"
-
-    con_list = [c1, c2]
-    ips = [c1.primaryIpAddress, c2.primaryIpAddress]
-    # Create external Service
-    random_name = random_str()
-    ext_service_name = random_name.replace("-", "")
-
-    ext_service = client.create_externalService(
-        name=ext_service_name, environmentId=env.id, externalIpAddresses=ips)
-
-    ext_service = client.wait_success(ext_service)
-    assert ext_service.state == "inactive"
-
-    return env, service, ext_service, con_list
 
 
 def activate_environment_with_external_services(
@@ -527,68 +474,3 @@ def test_extservice_add_and_delete_ips(super_client, client):
                               con_list)
 
     delete_all(client, [env])
-
-
-def validate_external_service(super_client, service, ext_services,
-                              exposed_port, container_list,
-                              exclude_instance=None,
-                              exclude_instance_purged=False):
-    time.sleep(5)
-
-    containers = get_service_container_list(super_client, service)
-    assert len(containers) == service.scale
-    for container in containers:
-        print "Validation for container -" + str(container.name)
-        host = super_client.by_id('host', container.hosts[0].id)
-        for ext_service in ext_services:
-            expected_dns_list = []
-            expected_link_response = []
-            dns_response = []
-            for con in container_list:
-                if (exclude_instance is not None) \
-                        and (con.id == exclude_instance.id):
-                    print "Excluded from DNS and wget list:" + con.name
-                else:
-                    expected_dns_list.append(con.primaryIpAddress)
-                    expected_link_response.append(con.externalId[:12])
-
-            print "Expected dig response List" + str(expected_dns_list)
-            print "Expected wget response List" + str(expected_link_response)
-
-            # Validate port mapping
-            ssh = paramiko.SSHClient()
-            ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
-            ssh.connect(host.ipAddresses()[0].address, username="root",
-                        password="root", port=int(exposed_port))
-
-            # Validate link containers
-            cmd = "wget -O result.txt --timeout=20 --tries=1 http://" + \
-                  ext_service.name + ":80/name.html;cat result.txt"
-            print cmd
-            stdin, stdout, stderr = ssh.exec_command(cmd)
-
-            response = stdout.readlines()
-            assert len(response) == 1
-            resp = response[0].strip("\n")
-            print "Actual wget Response" + str(resp)
-            assert resp in (expected_link_response)
-
-            # Validate DNS resolution using dig
-            cmd = "dig " + ext_service.name + " +short"
-            print cmd
-            stdin, stdout, stderr = ssh.exec_command(cmd)
-
-            response = stdout.readlines()
-            print "Actual dig Response" + str(response)
-
-            expected_entries_dig = len(container_list)
-            if exclude_instance is not None:
-                expected_entries_dig = expected_entries_dig - 1
-
-            assert len(response) == expected_entries_dig
-
-            for resp in response:
-                dns_response.append(resp.strip("\n"))
-
-            for address in expected_dns_list:
-                assert address in dns_response
