+++ tests/validation/cattlevalidationtest/core/test_services_links.py
@@ -1,48 +1,6 @@
 from common_fixtures import *  # NOQA
 
-WEB_IMAGE_UUID = "docker:sangeetha/testlbsd:latest"
-SSH_IMAGE_UUID = "docker:sangeetha/testclient:latest"
-
 logger = logging.getLogger(__name__)
-
-
-def create_env_with_2_svc(client, scale_svc, scale_consumed_svc, port):
-
-    launch_config_svc = {"imageUuid": SSH_IMAGE_UUID,
-                         "ports": [port+":22/tcp"]}
-
-    launch_config_consumed_svc = {"imageUuid": WEB_IMAGE_UUID}
-
-    # Create Environment
-    random_name = random_str()
-    env_name = random_name.replace("-", "")
-    env = client.create_environment(name=env_name)
-    env = client.wait_success(env)
-    assert env.state == "active"
-
-    # Create Service
-    random_name = random_str()
-    service_name = random_name.replace("-", "")
-    service = client.create_service(name=service_name,
-                                    environmentId=env.id,
-                                    launchConfig=launch_config_svc,
-                                    scale=scale_svc)
-
-    service = client.wait_success(service)
-    assert service.state == "inactive"
-
-    # Create Consumed Service
-    random_name = random_str()
-    service_name = random_name.replace("-", "")
-
-    consumed_service = client.create_service(
-        name=service_name, environmentId=env.id,
-        launchConfig=launch_config_consumed_svc, scale=scale_consumed_svc)
-
-    consumed_service = client.wait_success(consumed_service)
-    assert consumed_service.state == "inactive"
-
-    return env, service, consumed_service
 
 
 def create_environment_with_linked_services(
@@ -648,75 +606,3 @@ def test_link_services_delete_instance(super_client, client):
     validate_linked_service(super_client, service, [consumed_service], port)
 
     delete_all(client, [env])
-
-
-def validate_linked_service(super_client, service, consumed_services,
-                            exposed_port, exclude_instance=None,
-                            exclude_instance_purged=False):
-    time.sleep(5)
-
-    containers = get_service_container_list(super_client, service)
-    assert len(containers) == service.scale
-
-    for con in containers:
-        host = super_client.by_id('host', con.hosts[0].id)
-        for consumed_service in consumed_services:
-            expected_dns_list = []
-            expected_link_response = []
-            dns_response = []
-            containers = get_service_container_list(super_client,
-                                                    consumed_service)
-            if exclude_instance_purged:
-                assert len(containers) == consumed_service.scale - 1
-            else:
-                assert len(containers) == consumed_service.scale
-
-            for con in containers:
-                if (exclude_instance is not None) \
-                        and (con.id == exclude_instance.id):
-                    logger.info("Excluded from DNS and wget list:" + con.name)
-                else:
-                    expected_dns_list.append(con.primaryIpAddress)
-                    expected_link_response.append(con.externalId[:12])
-
-            logger.info("Expected dig response List" + str(expected_dns_list))
-            logger.info("Expected wget response List" +
-                        str(expected_link_response))
-
-            # Validate port mapping
-            ssh = paramiko.SSHClient()
-            ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
-            ssh.connect(host.ipAddresses()[0].address, username="root",
-                        password="root", port=int(exposed_port))
-
-            # Validate link containers
-            cmd = "wget -O result.txt http://" + consumed_service.name + \
-                  ":80/name.html;cat result.txt"
-            logger.info(cmd)
-            stdin, stdout, stderr = ssh.exec_command(cmd)
-
-            response = stdout.readlines()
-            assert len(response) == 1
-            resp = response[0].strip("\n")
-            logger.info("Actual wget Response" + str(resp))
-            assert resp in (expected_link_response)
-
-            # Validate DNS resolution using dig
-            cmd = "dig " + consumed_service.name + " +short"
-            logger.info(cmd)
-            stdin, stdout, stderr = ssh.exec_command(cmd)
-
-            response = stdout.readlines()
-            logger.info("Actual dig Response" + str(response))
-
-            expected_entries_dig = consumed_service.scale
-            if exclude_instance is not None:
-                expected_entries_dig = expected_entries_dig - 1
-
-            assert len(response) == expected_entries_dig
-
-            for resp in response:
-                dns_response.append(resp.strip("\n"))
-
-            for address in expected_dns_list:
-                assert address in dns_response
